refactor: add App.js to App folder
